feat: Log모듈 추가 및 연결

+++ src/modules/issues/utils/sync-calendar-event.ts
@@ -1,11 +1,20 @@
 // src/modules/issues/utils/sync-calendar-event.ts
-import { Issue } from "../entities/issue.entity";
+import { Issue } from "src/modules/issues/entities/issue.entity";
 import { CalendarService } from "../../calendar/calendar.service";
+import { Repository } from "typeorm";
 
-export async function syncCalendarEvent(issue: Issue, calendarService: CalendarService) {
+export async function syncCalendarEvent(
+  issue: Issue,
+  calendarService: CalendarService,
+  issueRepository: Repository<Issue>,
+) {
   if (!issue.startAt || !issue.endAt) {
+    if (issue.calendarEventId) {
     // 일정 정보 없으면 이벤트 삭제
-    await calendarService.deleteEventByIssue(issue);
+      await calendarService.deleteEventByIssue(issue);
+      issue.calendarEventId = null;
+      await issueRepository.save(issue);
+    }
     return;
   }
 
@@ -13,6 +22,7 @@ export async function syncCalendarEvent(issue: Issue, calendarService: CalendarS
     // 새 이벤트 생성
     const event = await calendarService.createEventFromIssue(issue);
     issue.calendarEventId = event.id;
+    await issueRepository.save(issue);
   } else {
     // 기존 이벤트 업데이트
     await calendarService.updateEventFromIssue(issue);
